requirements.txt
playwright
pytest
google-generativeai
python-dotenv
